Updated the Makefile Again, now you can specify the following directories to install into
BASEDIR The base of the install path all other directories are relative to this
BINDIR Base Directory to install binaries into, relative to BASEDIR
CONFDIR Base Direcotry to install configs into, relative to BASEDIR

Edited the Debian init.d scripts to read as much of their config as possible from their
respective config files.

Update the makefile with an install-debian target which installs the init scripts and
correctly sets the paths to the installed programs and config files in them.