Rurallink needs a L2 tunnel solutions to turn a L3 network of AP's into a L2 network so that clients of the AP can roam nicly between the AP's.
Etud seemed like a great building block.

This commit adds:
 * Dynamic mac address learning
 * Dynamic endpoint learning (for broadcast domain)
 * Unknown mac address forwarding (via static/WANd peers)
 * Broadcast relaying (for when using as a tree rather than full mesh)

This commit has a focus slightly different from traditional WAND. RuralLink
needs a tree formation, not full mesh. Clients of the AP's cannot talk to
each other, only the one Etud with dynamic learning turned on will send
broadcast traffic that will be seen on all AP's.

In theory if unknown mac forwarding is turned on, dynamic learning is turned on
and WANd is used to create a full mesh, sane things will happen, but its not that
tested.